refactor(core): rename the dynamic module to manifest_build
The Python introspection layer was removed in this branch; the module now
only holds static manifest-building (rebuild_manifest_full) and third-party
hashing (compute_third_party_hash). Rename to reflect reality — nothing
'dynamic' remains.
