+++ crates/toolr-core/src/manifest_build/hash.rs
