+++ crates/toolr-core/src/manifest_build/rebuild.rs
